adds renedered html file
Summary:
This change adds the rendered shieldfnt ingestion API. The source is in:
https://github.com/moloco/marvel2/blob/master/api/shieldfnt-ingestion.apib

(It's out of sync, as we are working on this in D9770).

The rendered document on online can be found at:
https://shieldfntingestion.docs.apiary.io

The HTML page in this diff uses another rendered; so, the look is different but the contents are equivalent.

Test Plan: none required.

Reviewers: hyeonseo, jeon

Subscribers: all

Differential Revision: https://moloco.phacility.com/D9781